+++ src/wifi/test/spectrum-wifi-phy-test.cc
@@ -28,10 +28,10 @@
 
 using namespace ns3;
 
-static const uint16_t CHANNEL_NUMBER = 36;
+static const uint8_t CHANNEL_NUMBER = 36;
 static const uint32_t FREQUENCY = 5180; // MHz
-static const uint32_t CHANNEL_WIDTH = 20; // MHz
-static const uint32_t GUARD_WIDTH = 10; // MHz
+static const uint8_t CHANNEL_WIDTH = 20; // MHz
+static const uint8_t GUARD_WIDTH = 10; // MHz
 
 /**
  * \ingroup wifi-test
